backend/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.0.0
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.0
djangorestframework-simplejwt==5.3.0
python-decouple==3.8
psycopg2-binary==2.9.6
celery==5.3.1
redis==4.5.5